requirements-webhook.txt
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
